Add new feature and Completion
Completion of the basic functions of the main game.
Added an enemy health bar and a script to control the length of the health bar.
Added a script for delegating when an enemy is destroyed, so that enemies can let other objects know when they are destroyed.
Added different types of bullets, bullet behavior data, and a script for the bullet firing function.
Added the remaining distance of enemy movement to determine which enemy is closest to the end point.
Increased the monster's detection range (attack range) and triggers for enemies, and also increased the detection range of enemies. When an enemy enters the monster's detection range, the monster will add the enemy to its list, and remove it from the list when it leaves the range.
Made prefabricated parts for enemy 2 (blue bug) and 3 types of bullets.